Exec

+++ src/utils/free/free_utils.c
@@ -6,13 +6,47 @@
 /*   By: mniemaz <mniemaz@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/22 18:27:41 by mniemaz           #+#    #+#             */
-/*   Updated: 2025/05/26 21:19:41 by mniemaz          ###   ########.fr       */
+/*   Updated: 2025/05/27 11:42:14 by mniemaz          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-void	ft_free_redirects(t_redirect **redirects)
+// void	ft_free_redirects(t_redirect **redirects)
+// {
+// 	int	i;
+
+// 	if (!redirects)
+// 		return ;
+// 	i = 0;
+// 	while (redirects[i])
+// 	{
+// 		if (redirects[i]->redir_type == HEREDOC)
+// 			unlink(redirects[i]->filename);
+// 		if (redirects[i]->filename != NULL)
+// 			free(redirects[i]->filename);
+// 		free(redirects[i]);
+// 		i++;
+// 	}
+// 	free(redirects);
+// }
+
+// void	ft_free_cmd_content(void *content)
+// {
+// 	t_cmd	*cmd;
+
+// 	if (!content)
+// 		return ;
+// 	cmd = cast_to_cmd(content);
+// 	if (cmd->args != NULL)
+// 		ft_free_tab((void **)cmd->args);
+// 	if (cmd->redirects != NULL)
+// 		ft_free_redirects(cmd->redirects);
+// 	free(cmd);
+// }
+
+
+void	ft_free_redirects(t_context *ctx, t_redirect **redirects)
 {
 	int	i;
 
@@ -21,7 +55,7 @@ void	ft_free_redirects(t_redirect **redirects)
 	i = 0;
 	while (redirects[i])
 	{
-		if (redirects[i]->redir_type == HEREDOC)
+		if (redirects[i]->redir_type == HEREDOC && ctx->hd_pid > 0)
 			unlink(redirects[i]->filename);
 		if (redirects[i]->filename != NULL)
 			free(redirects[i]->filename);
@@ -31,7 +65,7 @@ void	ft_free_redirects(t_redirect **redirects)
 	free(redirects);
 }
 
-void	ft_free_cmd_content(void *content)
+void	ft_free_cmd_content_ctx(t_context *ctx, void *content)
 {
 	t_cmd	*cmd;
 
@@ -41,8 +75,23 @@ void	ft_free_cmd_content(void *content)
 	if (cmd->args != NULL)
 		ft_free_tab((void **)cmd->args);
 	if (cmd->redirects != NULL)
-		ft_free_redirects(cmd->redirects);
+		ft_free_redirects(ctx, cmd->redirects);
 	free(cmd);
+}
+
+void ft_free_cmds_content_ctx(t_context *ctx)
+{
+	t_node *curr;
+	t_node *next;
+
+	curr = *ctx->head_cmd;
+	while(curr)
+	{
+		next = curr->next;
+		ft_free_cmd_content_ctx(ctx, curr->content);
+		free(curr);
+		curr = next;
+	}
 }
 
 void	free_exec(t_exec *data)
@@ -64,8 +113,10 @@ void	ft_free_ctx_cmds(t_context *context)
 	if (context->head_cmd == NULL)
 		return ;
 	close_all_cmds_redirs(context->head_cmd);
+	// if (*context->head_cmd)
+	// 	ft_lstclear(context->head_cmd, ft_free_cmd_content);
 	if (*context->head_cmd)
-		ft_lstclear(context->head_cmd, ft_free_cmd_content);
+		ft_free_cmds_content_ctx(context);
 	free(context->head_cmd);
 	context->head_cmd = NULL;
 }
